implement course route
