+++ src/app/api/instructors/route.js
@@ -1,39 +1,20 @@
 import { NextResponse } from 'next/server';
 import clientPromise from '../../../../mongodb';
 
-// Named export for the GET HTTP method
-export async function GET(req) {
+// Fetch the instructors from the 'niner-rate.instructors' collection
+export async function GET(_req) {
   try {
-    // Extract the course ID from the request URL
-    // Note: Adjust the way you access `id` based on how your URL parameters are being passed.
-    const id = req.url.searchParams.get('id');
-    console.log('ID from URL:', id);
-
-    // Convert the ID from a string to an ObjectId for MongoDB querying
-    const objectId = new ObjectId(id);
-
-    // Connect to the database
     const client = await clientPromise;
     const db = client.db('niner-rate');
 
-    // Query for the specific course by its ObjectId
-    const course = await db.collection('courses').findOne({ _id: objectId });
+    // Since the instructors are stored in an array within a single document,
+    // you need to first fetch the document and then access the professors array.
+    const data = await db.collection('instructors').findOne({});
+    const professors = data.professors || [];
 
-    // If the course doesn't exist, return a 404 response
-    if (!course) {
-      return new Response('Course not found', { status: 404 });
-    }
-
-    // Return the course data
-    return new Response(JSON.stringify(course), {
-      status: 200,
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    });
+    return NextResponse.json(professors); // Use NextResponse to return JSON data
   } catch (error) {
-    // Log and return the error
-    console.error('Failed to fetch course:', error);
-    return new Response('Failed to fetch course', { status: 500 });
+    console.error('Failed to fetch professors:', error);
+    return new NextResponse('Failed to fetch professors', { status: 500 }); // Use NextResponse to return an error
   }
 }
